Fix several warnings
* Avoid many implicit type conversions or make them at least explicit
    * Use range-based for loops instead of an explicit index where possible
      to eliminate the problematic variable/comparison completely
    * Use float literals to avoid implicit conversions to double (and then
      back to float)
    * Replace some macros with constexpr functions that return the correct
      type
* Port away from deprecated FluidSynth and Qt classes and functions, most
  notably:
    * Use QElapsedTimer instead of QTime and change the type used to store
      ticks to qint64 accordingly
* Declare unused functions parameters as such explicitly to silence related
  warnings

+++ src/Chord.h
@@ -196,8 +196,8 @@ public:
         m_cordSpanGapTime = 0;
         m_completeChord.clear();
         m_currentChord.clear();
-        m_cfg_ChordNoteGap = CMidiFile::ppqnAdjust(Cfg::chordNoteGap());
-        m_cfg_ChordMaxLength = CMidiFile::ppqnAdjust(Cfg::chordMaxLength());
+        m_cfg_ChordNoteGap = CMidiFile::ppqnAdjust(static_cast<float>(Cfg::chordNoteGap()));
+        m_cfg_ChordMaxLength = CMidiFile::ppqnAdjust(static_cast<float>(Cfg::chordMaxLength()));
     }
 
     CChord getChord()
